enable `rate` support

+++ src/v0.9/fastq_filter.h
@@ -89,7 +89,7 @@ void fastF(gzFile file_in[3],
            gzFile file_out[3],
            node *tree_whitelist,
            unsigned int seed,
-           double rate);
+           float rate);
 
 // int main() {
 //     struct queue *q = init_queue();
